# commercetoolsDemo/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /me/order:
    get:
      tags:
      - me-controller
      operationId: getMyOrders
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      tags:
      - me-controller
      operationId: createMyOrder
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /me/cart:
    get:
      tags:
      - me-controller
      operationId: getMyCarts
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartListResponse"
    post:
      tags:
      - me-controller
      operationId: createMyCart
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCartRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /me/cart/{id}/update:
    post:
      tags:
      - me-controller
      operationId: updateMyCart
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /me/cart/{id}/shipping-method:
    post:
      tags:
      - me-controller
      operationId: setShippingMethod
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetShippingMethodRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /me/cart/{id}/shipping-address:
    post:
      tags:
      - me-controller
      operationId: setShippingAddress
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /me/cart/{id}/billing-address:
    post:
      tags:
      - me-controller
      operationId: setBillingAddress
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /me/cart/{id}/add-item:
    post:
      tags:
      - me-controller
      operationId: addLineItemToCart
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLineItemRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /auth/register:
    post:
      tags:
      - auth-controller
      operationId: registerCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CustomerResponse"
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: loginCustomer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /admin/cart:
    post:
      tags:
      - admin-controller
      operationId: createCart
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCartRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /me/order/{id}:
    get:
      tags:
      - me-controller
      operationId: getMyOrderById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /me/cart/{id}:
    get:
      tags:
      - me-controller
      operationId: getMyCartById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
    delete:
      tags:
      - me-controller
      operationId: deleteMyCart
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /me/cart/active:
    get:
      tags:
      - me-controller
      operationId: getMyActiveCart
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
  /auth/me:
    get:
      tags:
      - auth-controller
      operationId: getCustomerInfo
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CustomerResponse"
  /admin/cart/{id}:
    get:
      tags:
      - admin-controller
      operationId: getCart
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
    delete:
      tags:
      - admin-controller
      operationId: deleteCart
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CartResponse"
components:
  schemas:
    CreateOrderRequest:
      type: object
      properties:
        cartId:
          type: string
        cartVersion:
          type: integer
          format: int64
    AddressResponse:
      type: object
      properties:
        streetName:
          type: string
        streetNumber:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    LineItemResponse:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        name:
          type: string
        quantity:
          type: integer
          format: int32
        price:
          $ref: "#/components/schemas/MoneyResponse"
        totalPrice:
          $ref: "#/components/schemas/MoneyResponse"
        productSlug:
          type: string
        imageUrls:
          type: array
          items:
            type: string
    MoneyResponse:
      type: object
      properties:
        currencyCode:
          type: string
        amount:
          type: number
    OrderResponse:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        orderState:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItemResponse"
        totalPrice:
          $ref: "#/components/schemas/MoneyResponse"
        shippingAddress:
          $ref: "#/components/schemas/AddressResponse"
        shippingInfo:
          $ref: "#/components/schemas/ShippingInfoResponse"
    ShippingInfoResponse:
      type: object
      properties:
        shippingMethodName:
          type: string
        price:
          $ref: "#/components/schemas/MoneyResponse"
    CreateCartRequest:
      type: object
      properties:
        currency:
          type: string
        anonymousId:
          type: string
    CartResponse:
      type: object
      properties:
        id:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItemResponse"
        totalPrice:
          $ref: "#/components/schemas/MoneyResponse"
        taxedPrice:
          $ref: "#/components/schemas/MoneyResponse"
        shippingAddress:
          $ref: "#/components/schemas/AddressResponse"
        billingAddress:
          $ref: "#/components/schemas/AddressResponse"
        shippingInfo:
          $ref: "#/components/schemas/ShippingInfoResponse"
    Address:
      type: object
      properties:
        streetName:
          type: string
        streetNumber:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
    CartAction:
      type: object
      properties:
        action:
          type: string
        productId:
          type: string
        variantId:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        address:
          $ref: "#/components/schemas/Address"
        shippingMethod:
          $ref: "#/components/schemas/ShippingMethodReference"
    CartUpdateRequest:
      type: object
      properties:
        version:
          type: integer
          format: int64
        actions:
          type: array
          items:
            $ref: "#/components/schemas/CartAction"
    ShippingMethodReference:
      type: object
      properties:
        typeId:
          type: string
        id:
          type: string
    SetShippingMethodRequest:
      type: object
      properties:
        version:
          type: integer
          format: int64
        shippingMethodId:
          type: string
    AddressRequest:
      type: object
      properties:
        version:
          type: integer
          format: int64
        address:
          $ref: "#/components/schemas/Address"
    AddLineItemRequest:
      type: object
      properties:
        version:
          type: integer
          format: int64
        productId:
          type: string
        variantId:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
    CreateCustomerRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    CustomerResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int32
        customer:
          $ref: "#/components/schemas/CustomerResponse"
    OrderListResponse:
      type: object
      properties:
        total:
          type: integer
          format: int32
        results:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
    CartListResponse:
      type: object
      properties:
        total:
          type: integer
          format: int32
        results:
          type: array
          items:
            $ref: "#/components/schemas/CartResponse"